type 관리 방식 변경
